requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
pydantic
openai
redis>=4.6.0
